feat: lambda building added
