--- cdk/.env.development
@@ -1,2 +1,0 @@
-PRODUCTS_TABLE_NAME=products
-STOCS_TABLE_NAME=stocs
