+++ cdk/package.json
@@ -8,20 +8,25 @@
     "build": "tsc",
     "watch": "tsc -w",
     "test": "jest",
-    "cdk": "cdk"
+    "cdk": "cdk",
+    "deploy": "cdk deploy --profile admin",
+    "destroy": "cdk destroy -f --profile admin",
+    "lambda": "ts-node bin/lambda-build.ts"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
     "@types/node": "20.12.7",
+    "aws-cdk": "2.145.0",
+    "esbuild": "0.21.5",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.2",
-    "aws-cdk": "2.145.0",
     "ts-node": "^10.9.2",
     "typescript": "~5.4.5"
   },
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.602.0",
     "@aws-sdk/lib-dynamodb": "^3.602.0",
+    "@types/aws-lambda": "^8.10.140",
     "aws-cdk-lib": "2.145.0",
     "constructs": "^10.0.0",
     "source-map-support": "^0.5.21",
