+++ cdk/services/serverless/product-service/GetProductsList.ts
@@ -1,5 +1,5 @@
 import { Construct } from 'constructs';
-import { IFunction } from 'aws-cdk-lib/aws-lambda';
+import { Code, IFunction } from 'aws-cdk-lib/aws-lambda';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 //import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
 //import { Runtime } from 'aws-cdk-lib/aws-lambda';
@@ -15,13 +15,27 @@ export class GetProductsList {
         //const getProductsList = new NodejsFunction(this, 'getProductsListLambda', {
         //    entry: 'services/serverless/product-service/functions/cdk-products.js',
         //    runtime: Runtime.NODEJS_LATEST,
-        //});
-
+        //});4
+/*        
+        const layer = new lambda.LayerVersion( 
+            scope,
+            'Layer_'+id,
+            {
+                code: lambda.Code.fromAsset('./lib/layers'),
+                compatibleRuntimes: [lambda.Runtime.NODEJS_20_X],
+                layerVersionName: 'NodeJSLayer'
+            }
+        )
+*/
         // Define the Lambda function resource
         const getProductsList = new lambda.Function(scope, id, {
-            runtime: lambda.Runtime.NODEJS_16_X, // Choose any supported Node.js runtime
+            runtime: lambda.Runtime.NODEJS_20_X, // Choose any supported Node.js runtime
             code: lambda.Code.fromAsset('services/serverless/product-service/functions'), // Points to the lambda directory
             handler: 'cdk-products.handler', // Points to the 'cdk-products' file in the lambda directory
+//            layers: [
+//                layer
+//            ],
+            
         });
 
         this.getProductsList = getProductsList;
