+++ cdk/services/serverless/product-service/functions/cdk-product.ts
@@ -1,16 +1,12 @@
-import { DynamoDB } from '@aws-sdk/client-dynamodb';
-import { ScanCommand } from '@aws-sdk/lib-dynamodb';
 
-const dynamodb = new DynamoDB({});
+import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda';
 
-exports.handler = async (event) => {
+export const handler: APIGatewayProxyHandler = async (
+    event: APIGatewayProxyEvent,
+    ): Promise<APIGatewayProxyResult> => {
 
-    const result = await dynamodb.send(
-        new ScanCommand({
-            TableName: process.env.PRODUCTS_TABLE_NAME,
-        })
-    );
-/*
+    console.log('Product lambda event: ', JSON.stringify(event, null, 4));
+
     const products = [
         {
             description: "Short Product Description1",
@@ -49,15 +45,32 @@ exports.handler = async (event) => {
             title: "ProductName",
         },
     ];
-*/
 
-    return {
-        statusCode: 200,
-        headers: { 
-            "Content-Type":"application/json",
-            'Access-Control-Allow-Origin': '*',
-            'Access-Control-Allow-Methods': 'OPTIONS,GET'
-        },
-        body: JSON.stringify(result.Items),
-    };
+    const productId = event.pathParameters?.product_id
+    const product = products.find((el) => el.id == productId)
+    
+    if (product) {
+        return {
+            statusCode: 200,
+            headers: { 
+                "Content-Type":"application/json",
+                'Access-Control-Allow-Origin': '*',
+                'Access-Control-Allow-Methods': 'OPTIONS,GET,POST,PUT,DELETE'
+            },
+            body: JSON.stringify(product),
+        };
+    
+    } else {
+        return {
+            statusCode: 404,
+            headers: { 
+                "Content-Type":"application/json",
+                'Access-Control-Allow-Origin': '*',
+                'Access-Control-Allow-Methods': 'OPTIONS,GET,POST,PUT,DELETE'
+            },
+            body: JSON.stringify({"message": "Product not found"}),
+        };
+
+    }
+
 };
