+++ cdk/tsconfig.json
@@ -8,7 +8,14 @@
     ],
     "declaration": true,
     "strict": true,
+    "preserveConstEnums": true,
+    "noEmit": true,
+    "sourceMap": false,
     "noImplicitAny": true,
+    "moduleResolution":"node",
+    "esModuleInterop": true, 
+    "forceConsistentCasingInFileNames": true, 
+    "isolatedModules": true, 
     "strictNullChecks": true,
     "noImplicitThis": true,
     "alwaysStrict": true,
@@ -26,6 +33,8 @@
   },
   "exclude": [
     "node_modules",
-    "cdk.out"
+    "cdk.out",
+    "**/*.test.ts",
+    "dist"
   ]
 }
